[fix] 위치 입력 드롭박스 모듈화 해제.
모듈화 하면 작동 안함.
또한 무리한 모듈화라 판단.
